Move DisplayAlert/ActionSheet code to a helper class so embedded Forms can use it
